fix(#652): Avoid actionable tooltips to get unreadable
This fix prevents some bad rendering of some actionable tooltips (i.e.
when the app asks if a todo with all tasks just marked as done shall be
closed as well) - on devies with too little horizontal screen estate,
the rendering would break and the label would be spread too much
vertically.