refactor: Continued project structure refactoring
This is the beginning of some refactoring of the project to aid bringing
it up to date with the current way of developing Qt/QML applications.
Concrete, this:

- Unifies all sources in the `src` folder.
- Uses Qt CMake functions to bind QML code into the project.

This enables some cool features, like better and more reliable
QML linting and use of the QML language server (qmlls), such that the
project can easily be edited in e.g. VSCode.

Further adaptions need to be made to finish this, such as declaratively
marking classes for use in QML.